refactor : entity 및 dto 생성자 정리
