+++ src/main/java/com/game/algo/algo/dto/request/BlockDraw.java
@@ -1,4 +1,4 @@
-package com.game.algo.algo.dto;
+package com.game.algo.algo.dto.request;
 
 import com.game.algo.algo.data.BlockColor;
 import lombok.AllArgsConstructor;
@@ -8,13 +8,11 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class JokerUpdate {
+public class BlockDraw {
 
     private Long playerId;
 
     private Long gameRoomId;
 
     private BlockColor blockColor;
-
-    private Integer index;
 }
