+++ src/main/java/com/game/algo/algo/dto/request/GameRoomCreate.java
@@ -1,4 +1,4 @@
-package com.game.algo.algo.dto;
+package com.game.algo.algo.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,11 +7,9 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class PlayerReadyUpdate {
+public class GameRoomCreate {
 
     private Long playerId;
 
-    private Long gameRoomId;
-
-    private Boolean ready;
+    private String title;
 }
