+++ src/main/java/com/game/algo/algo/dto/request/GameRoomJoin.java
@@ -1,4 +1,4 @@
-package com.game.algo.algo.dto;
+package com.game.algo.algo.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,13 +7,9 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class StartBlockDraw {
-
-    private Long playerId;
+public class GameRoomJoin {
 
     private Long gameRoomId;
 
-    private int whiteBlockCount = 0;
-
-    private int blackBlockCount = 0;
+    private Long playerId;
 }
