+++ src/main/java/com/game/algo/algo/dto/request/GameStart.java
@@ -1,4 +1,4 @@
-package com.game.algo.algo.dto.messagetype;
+package com.game.algo.algo.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,9 +7,8 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class PlayerCreate {
+public class GameStart {
 
-    private String name;
-
-    private String sessionId;
+    private Long gameRoomId;
+    private Long playerId;
 }
