+++ src/main/java/com/game/algo/algo/dto/request/JokerUpdate.java
@@ -1,0 +1,20 @@
+package com.game.algo.algo.dto.request;
+
+import com.game.algo.algo.data.BlockColor;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class JokerUpdate {
+
+    private Long playerId;
+
+    private Long gameRoomId;
+
+    private BlockColor blockColor;
+
+    private Integer index;
+}
