+++ src/main/java/com/game/algo/algo/dto/request/NextPhase.java
@@ -1,5 +1,6 @@
-package com.game.algo.algo.dto.messagetype;
+package com.game.algo.algo.dto.request;
 
+import com.game.algo.algo.entity.GameRoom;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -7,9 +8,13 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class GameRoomJoin {
+public class NextPhase {
 
     private Long gameRoomId;
 
     private Long playerId;
+
+    private GameRoom.Phase phase;
+
+    private Integer progressPlayerNum;
 }
