+++ src/main/java/com/game/algo/algo/dto/request/PlayerReadyUpdate.java
@@ -1,4 +1,4 @@
-package com.game.algo.algo.dto;
+package com.game.algo.algo.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,8 +7,11 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class GameStart {
+public class PlayerReadyUpdate {
+
+    private Long playerId;
 
     private Long gameRoomId;
-    private Long playerId;
+
+    private Boolean ready;
 }
