+++ src/main/java/com/game/algo/algo/dto/request/StartBlockDraw.java
@@ -1,6 +1,5 @@
-package com.game.algo.algo.dto;
+package com.game.algo.algo.dto.request;
 
-import com.game.algo.algo.data.BlockColor;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -8,11 +7,13 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class BlockDraw {
+public class StartBlockDraw {
 
     private Long playerId;
 
     private Long gameRoomId;
 
-    private BlockColor blockColor;
+    private int whiteBlockCount = 0;
+
+    private int blackBlockCount = 0;
 }
