+++ src/main/java/com/game/algo/websocket/aspect/WebSocketAspect.java
@@ -1,9 +1,6 @@
 package com.game.algo.websocket.aspect;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.game.algo.algo.dto.GameStatusData;
-import com.game.algo.algo.entity.GameRoom;
-import com.game.algo.websocket.data.MessageType;
 import com.game.algo.websocket.dto.MessageDataResponse;
 import lombok.RequiredArgsConstructor;
 import org.aspectj.lang.ProceedingJoinPoint;
