+++ src/main/java/com/game/algo/websocket/handler/WebSocketHandler.java
@@ -3,10 +3,7 @@ package com.game.algo.websocket.handler;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.game.algo.algo.controller.GameWebSocketMessageController;
 import com.game.algo.algo.data.GameProperty;
-import com.game.algo.algo.dto.*;
-import com.game.algo.algo.dto.messagetype.GameRoomCreate;
-import com.game.algo.algo.dto.messagetype.GameRoomJoin;
-import com.game.algo.algo.dto.messagetype.PlayerCreate;
+import com.game.algo.algo.dto.request.*;
 import com.game.algo.algo.exception.GameLogicException;
 import com.game.algo.algo.exception.StopMethodException;
 import com.game.algo.websocket.data.MessageType;
